requirements.txt
discord.py>=2.4.0,<3.0.0
requests>=2.31.0,<3.0.0
